+++ src/engine/economy/index.ts
@@ -5,3 +5,4 @@
 export * from './trade';
 export * from './playerToken';
 export * from './bank';
+export * from './cashSwipe';